refactor(action-controller): rename railtie.ts → trailtie.ts per trails convention
Trails railties are not Rails::Railtie subclasses (different lifecycle,
different surface). The activerecord precedent already encodes this
(`activerecord:railtie.rb` → `trailtie.ts` in
scripts/api-compare/conventions.ts:32). Extend the same mapping to
actioncontroller and rename the file.
